+++ components/collector/src/source_collectors/azure_devops/user_story_points.py
@@ -11,7 +11,7 @@ class AzureDevopsUserStoryPoints(AzureDevopsIssues):
     """Collector to get user story points from Azure Devops Server."""
 
     def _item_select_fields(self) -> list[str]:
-        """Also request date fields to calculate lead time."""
+        """Also request all work item fields which may hold story point values."""
         base_fields = super()._item_select_fields()
         return base_fields + ["Microsoft.VSTS.Scheduling.StoryPoints", "Microsoft.VSTS.Scheduling.Effort"]
 
@@ -33,7 +33,7 @@ class AzureDevopsUserStoryPoints(AzureDevopsIssues):
     async def _parse_value(self, responses: SourceResponses) -> Value:
         """Override to parse the sum of the user story points from the responses."""
         calculated_value = sum(self.__story_points(work_item) for work_item in await self._work_items(responses))
-        return str(decimal_round_half_up(calculated_value))  # follow frontend behaviour
+        return str(decimal_round_half_up(calculated_value))
 
     @staticmethod
     def __story_points(work_item: dict[str, dict[str, None | float]]) -> float: