+++ components/collector/src/source_collectors/jira/lead_time_for_changes.py
@@ -16,23 +16,20 @@ class JiraLeadTimeForChanges(JiraIssues):
     """Jira collector for lead time for changes."""
 
     def _include_issue(self, issue: dict) -> bool:
-        if not issue["fields"]["status"]["statusCategory"]["key"] == "done":
+        """Return whether this issue should be counted."""
+        if issue["fields"]["status"]["statusCategory"]["key"] != "done":
             return False
-
-        finished_date = issue["fields"].get("updated")  # TODO - how to find out statusCategory update ?
-        if not finished_date:
+        if not (finished_date := issue["fields"].get("updated")):
             return False
-
         return days_ago(parse(finished_date)) <= int(cast(str, self._parameter("lookback_days")))
 
     @classmethod
     def _compute_value(cls, entities: Entities) -> Value:
+        """Calculate the average lead time of the completed issues."""
         if not entities:
             return None
-
         lead_times = []
         for issue in entities:
             issue_lead_time = parse(issue["updated"]) - parse(issue["created"])
             lead_times.append(issue_lead_time.days)
-
         return str(round(mean(lead_times)))